Integrated one/requirements.txt
# requirements.txt - Trinetra AI Security Platform

# Core Backend
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0

# LLM & AI
langchain-ollama>=0.0.1
torch>=2.0.0
transformers>=4.35.0

# Web Scraping & Research
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.0

# Preprocessing Module
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
pyzxing>=0.4.0

# Testing
pytest>=7.0.0
